cleanup testing a bit
